Rename rpng_common.h to rpng_internal.h
