+++ libretro-common/formats/png/rpng_internal.h
