+++ libretro-common/formats/png/rpng_nbio.c
@@ -24,7 +24,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-#include "rpng_common.h"
+#include "rpng_internal.h"
 #include "rpng_decode.h"
 
 static bool read_chunk_header(uint8_t *buf, struct png_chunk *chunk)
